refactor(): meet package standards
Closes #1
