+++ .eslintrc
@@ -1,9 +1,11 @@
 {
   "extends": "eslint-config-cycle",
-  "ecmaFeatures": {
-    "jsx": true
+  "env": {
+    "browser": true,
+    "node": true
   },
   "rules": {
+    "new-cap": 0,
     "semi": [2, "always"]
   }
 }
