+++ .npmignore
@@ -1,0 +1,3 @@
+.idea/
+*.sublime-project
+*.sublime-workspace
