+++ .travis.yml
@@ -1,0 +1,3 @@
+language: node_js
+node_js:
+  - "4.1"
