+++ README.md
@@ -4,10 +4,25 @@
 
 ## Installation
 
+[![npm version](https://badge.fury.io/js/%40cyclic%2Futil-combine-class-names.svg)](http://badge.fury.io/js/%40cyclic%2Futil-combine-class-names)
+
 ```shell
 $ npm install @cyclic/util-combine-class-names
 ```
 
+### Usage
+
+See the [API documentation].
+
 ## License
 
-MIT © [Cyclic Materials](http://github.com/CyclicMaterials) 
+MIT © [Cyclic Materials](http://github.com/CyclicMaterials)
+
+- - -
+
+[![Build Status](https://travis-ci.org/CyclicMaterials/util-combine-class-names.svg)](https://travis-ci.org/CyclicMaterials/util-combine-class-names)
+[![Code Climate](https://codeclimate.com/github/CyclicMaterials/util-combine-class-names/badges/gpa.svg)](https://codeclimate.com/github/CyclicMaterials/util-combine-class-names)
+[![Dependency Status](https://david-dm.org/CyclicMaterials/util-combine-class-names.svg)](https://david-dm.org/CyclicMaterials/util-combine-class-names)
+[![devDependency Status](https://david-dm.org/CyclicMaterials/util-combine-class-names/dev-status.svg)](https://david-dm.org/CyclicMaterials/util-combine-class-names#info=devDependencies)
+
+[API documentation]: ./doc/api.md
